* 'feature/Q&A' of github.com:zhiyicx/thinksns-plus-android:
  (#1034) code fix this
  (#1045) code fix this
  (#1043) code fix this
  (#1027) code fix this
  (#1033) code fix this
  (#1022) code fix this
  (#987) code fix this
  (#881) code fix this
  (#872) code fix this
  (#975) code 提示信息修改

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/dynamic/send/SendDynamicFragment.java
@@ -430,7 +430,7 @@ public class SendDynamicFragment extends TSFragment<SendDynamicContract.Presente
                     SendDynamicDataBeanV2.StorageTaskBean taskBean = new SendDynamicDataBeanV2.StorageTaskBean();
                     ImageBean imageBean = selectedPhotos.get(i);
                     photos.add(imageBean);
-                    taskBean.setAmount(imageBean.getToll_monye() > 0 ? PayConfig.realCurrencyYuan2Fen(imageBean.getToll_monye()) : null);
+                    taskBean.setAmount(imageBean.getToll_monye() > 0 ?(long) (PayConfig.realCurrencyYuan2Fen(imageBean.getToll_monye())) : null);
                     taskBean.setType(imageBean.getToll_monye() * imageBean.getToll_type() > 0
                             ? (imageBean.getToll_type() == LOOK_TOLL ? LOOK_TOLL_TYPE : DOWNLOAD_TOLL_TYPE) : null);
                     storage_task.add(taskBean);
@@ -450,7 +450,7 @@ public class SendDynamicFragment extends TSFragment<SendDynamicContract.Presente
                     SendDynamicDataBeanV2.StorageTaskBean taskBean = new SendDynamicDataBeanV2.StorageTaskBean();
                     ImageBean imageBean = selectedPhotos.get(i);
                     photos.add(imageBean);
-                    taskBean.setAmount(imageBean.getToll_monye() > 0 ? PayConfig.realCurrencyYuan2Fen(imageBean.getToll_monye()) : null);
+                    taskBean.setAmount(imageBean.getToll_monye() > 0 ?(long)( PayConfig.realCurrencyYuan2Fen(imageBean.getToll_monye())) : null);
                     taskBean.setType(imageBean.getToll_monye() * imageBean.getToll_type() > 0
                             ? (imageBean.getToll_type() == LOOK_TOLL ? LOOK_TOLL_TYPE : DOWNLOAD_TOLL_TYPE) : null);
                 }
